perf: `setInterval` 系処理をメモ化 Fixes #238
