+++ packages/frontend/src/hooks/useInterval.ts
@@ -1,0 +1,21 @@
+import { useEffect, useRef } from "react";
+
+const noop = () => {};
+
+export const useInterval = (callback: () => void, delay: number | null) => {
+  const savedCallback = useRef(noop);
+
+  useEffect(() => {
+    savedCallback.current = callback;
+  }, [callback]);
+
+  useEffect(() => {
+    function tick() {
+      savedCallback.current();
+    }
+    if (delay !== null) {
+      let id = setInterval(tick, delay);
+      return () => clearInterval(id);
+    }
+  }, [delay]);
+};
